requirements-training.txt
torch
torchaudio
torchvision
matplotlib
Pillow
onnx
onnxruntime
onnxscript